SODEXO-23 Error loading user when LDAP service is unavailable
In the case of the LDAP service unavailable, we still catch the exception and return null, which has the consequence of storing the user key in the nonExistantUserCache which is clearly wrong since it means that even when the service comes back lookups of the user will still fails.

To test this, disconnect the LDAP server temporarily, try to login with a LDAP user, reconnect the LDAP. Before this fix the user should fail all the time once the LDAP service is back, while after this fix it will only fail until the service is back.